requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
SQLAlchemy==2.0.30
PyMySQL==1.1.1
mysql-connector-python==8.0.33
python-dotenv==1.0.1
